+++ vantage-sql/src/postgres/vista/factory.rs
@@ -73,6 +73,7 @@ impl PostgresVistaFactory {
                 can_fetch_window: true,
                 can_traverse_to_record: true,
                 can_traverse_to_set: true,
+                can_traverse_in_columns: true,
                 ..VistaCapabilities::default()
             },
             metadata,