+++ vantage-sql/tests/sqlite.rs
@@ -24,6 +24,8 @@ mod sqlite {
     mod identifier;
     #[path = "2_insert.rs"]
     mod insert;
+    #[path = "8_implicit_references.rs"]
+    mod implicit_references;
     #[path = "5_invariants.rs"]
     mod invariants;
     #[path = "2_primitives.rs"]